move some non-framework items out of `chia.cmds.cmd_classes`
